pyproject.toml
[project]
name = "clothing-classifier"
version = "0.1.0"
description = "Clothing image classifier using ResNet and PyTorch Lightning"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dvc>=3.55.0",
    "dvc-gdrive>=3.0.0",
    "fire>=0.6.0",
    "hydra-core>=1.3.2",
    "kaggle>=1.8.3",
    "matplotlib>=3.8.0",
    "mlflow>=2.10.0",
    "omegaconf>=2.3.0",
    "onnx>=1.15.0",
    "onnxruntime>=1.23.2",
    "onnxscript>=0.5.7",
    "pandas>=2.1.0",
    "pillow>=10.0.0",
    "pytorch-lightning>=2.2.0",
    "scikit-learn>=1.4.0",
    "tensorboard>=2.20.0",
    "tensorboardx>=2.6.4",
    "torch>=2.2.0",
    "torchmetrics>=1.3.0",
    "torchvision>=0.17.0",
    "tritonclient[http]>=2.64.0",
]

[project.optional-dependencies]
triton = [
    "tritonclient[http]>=2.42.0",
]

[dependency-groups]
dev = [
    "black>=25.12.0",
    "flake8>=7.3.0",
    "isort>=7.0.0",
    "pre-commit>=4.5.1",
]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 88

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
